requirements.txt
amqp==5.3.1
annotated-types==0.7.0
apify-fingerprint-datapoints==0.7.0
asgiref==3.11.0
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
browserforge==1.2.3
cachetools==6.2.2
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crawlee==1.2.0
cron-descriptor==2.0.6
dj-database-url==3.0.1
django==5.1.4
django-celery-beat==2.8.1
django-celery-results==2.6.0
django-timezone-field==7.2.1
exceptiongroup==1.3.1
filelock==3.20.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.31.0
idna==3.11
impit==0.9.3
kombu==5.6.1
lxml==6.0.2
more-itertools==10.8.0
multidict==6.7.0
packaging==25.0
pillow==12.0.0
playwright==1.56.0
prompt-toolkit==3.0.52
propcache==0.4.1
protego==0.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyee==13.0.0
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==6.4.0
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
tldextract==5.3.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.1
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
whitenoise==6.7.0

